Feat: useGetRooms -> useGetData 파일명 변경
